Fixed a bug in the Time class to handle wrap around indexing a bit better
Added a blackout for areas with no tiles (needs more work)

Moved resolution to the game buffer (was previous set in the win32 wrapper), so that it is independent of window size, and the window will now scale the game to fit instead of trying to fill the area with more tiles.

Implemented a (hacky) solution to solve the issue of the game freezing whenever the window is resizing or moving. May need some love.